move DynamicWebsitePreviewFAQ in its own file
